jetson/requirements.txt
Flask==0.10.1
Flask-SocketIO==0.6.0
gevent==1.0.2
gevent-socketio==0.3.6
gevent-websocket==0.9.5
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.10.4
Pillow==3.1.1
pynetworktables==2015.3.0
python-engineio==0.8.7
python-socketio==1.0
six==1.10.0
Werkzeug==0.11.3
